refactor(engineering): keep the coding loop in this repo under coding-loop/
A second GitHub repository is not required. The reviewer already runs from
this tree. Decision [2026-08-17-0003] now records that.
