+++ tests/issue-coding-loop.test.js
@@ -95,11 +95,12 @@ test('architecture doc names the three identities and the plan cycle stop', () =
   assert.match(doc, /highest-generation preview/);
 });
 
-test('architecture does not name the NewPush Slack Mastra repo as the host', () => {
+test('architecture keeps the loop in this repo under coding-loop/', () => {
   const doc = read('docs/architecture/issue-coding-loop.md');
-  assert.match(doc, /not\*\* `newpush\/newpush-mastra-orchestration`/);
-  assert.match(doc, /scripts\/issue-loop\//);
-  assert.match(doc, /public `project-noemi` host repo/);
+  assert.match(doc, /newpush\/newpush-mastra-orchestration/);
+  assert.match(doc, /not a second GitHub repo/);
+  assert.match(doc, /coding-loop\//);
+  assert.ok(require('fs').existsSync(require('path').join(__dirname, '..', 'coding-loop', 'run.js')));
 });
 
 test('architecture and conductor treat GitHub 429/5xx as retries, not verdicts', () => {
